# refactor-baseline/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "wardenIQ \u2014 Test Intelligence Platform",
    "version": "0.1.0-beta"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/features/{fid}/export/csv": {
      "get": {
        "operationId": "feature_export_csv_api_features__fid__export_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feature Export Csv"
      }
    },
    "/api/features/{fid}/export/csv-selected": {
      "post": {
        "operationId": "feature_export_selected_csv_api_features__fid__export_csv_selected_post",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feature Export Selected Csv"
      }
    },
    "/api/features/{fid}/export/pdf": {
      "get": {
        "operationId": "feature_export_pdf_api_features__fid__export_pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feature Export Pdf"
      }
    },
    "/api/features/{fid}/export/pdf-selected": {
      "post": {
        "operationId": "feature_export_selected_pdf_api_features__fid__export_pdf_selected_post",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feature Export Selected Pdf"
      }
    },
    "/api/features/{fid}/gap/automation/export/{fmt}": {
      "get": {
        "operationId": "export_gap_automation_api_features__fid__gap_automation_export__fmt__get",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "fmt",
            "required": true,
            "schema": {
              "title": "Fmt",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Gap Automation"
      }
    },
    "/api/features/{fid}/gap/pr-coverage/export/{fmt}": {
      "get": {
        "operationId": "export_gap_pr_coverage_api_features__fid__gap_pr_coverage_export__fmt__get",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "fmt",
            "required": true,
            "schema": {
              "title": "Fmt",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Gap Pr Coverage"
      }
    },
...
    "/api/features/{fid}/report.pdf": {
      "get": {
        "operationId": "feature_report_api_features__fid__report_pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feature Report"
      }
    },
...
  }
}